refactor(xlf/renderer): hooks
